+++ packages/@cldcvr/flow-lineage/src/components/f-lineage/draw/draw-lineage.ts
@@ -5,5 +5,4 @@ import drawNodes from "./draw-nodes";
 export default function drawLineage(params: DrawLineageParams) {
   drawNodes(params);
   drawLinks(params);
-  console.log("Done plotting!");
 }